Trait txpool::Listener

pub trait Listener<T> {
    fn added(&mut self, _tx: &Arc<T>, _old: Option<&Arc<T>>) { ... }

    fn rejected<H: Debug + LowerHex>(
        &mut self, 
        _tx: &Arc<T>, 
        _reason: &Error<H>
    ) { ... }

    fn dropped(&mut self, _tx: &Arc<T>, _by: Option<&T>) { ... }

    fn invalid(&mut self, _tx: &Arc<T>) { ... }

    fn canceled(&mut self, _tx: &Arc<T>) { ... }

    fn culled(&mut self, _tx: &Arc<T>) { ... }
}

Transaction pool listener.

Listener is being notified about status of every transaction in the pool.

Provided methods

fn added(&mut self, _tx: &Arc<T>, _old: Option<&Arc<T>>)

The transaction has been successfuly added to the pool. If second argument is Some the transaction has took place of some other transaction which was already in pool. NOTE: You won’t be notified about drop of old transaction separately.

fn rejected<H: Debug + LowerHex>(&mut self, _tx: &Arc<T>, _reason: &Error<H>)

The transaction was rejected from the pool. It means that it was too cheap to replace any transaction already in the pool.

fn dropped(&mut self, _tx: &Arc<T>, _by: Option<&T>)

The transaction was pushed out from the pool because of the limit.

fn invalid(&mut self, _tx: &Arc<T>)

The transaction was marked as invalid by executor.

fn canceled(&mut self, _tx: &Arc<T>)

The transaction has been canceled.

fn culled(&mut self, _tx: &Arc<T>)

The transaction has been culled from the pool.